[IMP]custom_lab_report: Change in action_request function of lab.appointment
At base fucntion in val of request_line we have (6,0,ids) format
that replace the list of linked IDs in record i.e every time when
record is created the link is lost or replace so we need to change
this format to (0,0,{values}) that will link to a new that needs to
be created with the given values dictionary. So we overwrite the base
function of one2Many field value in table.