[IMP] add transaction and new copy of original policy
Previous Policy for Copy of Original:
The first print is original one and any prints after the
first one there needs to be "Copy of Original" and copy
count written in the invoice.

Current Policy for Copy of Original:
The first print must print "Tax Invoice" and the second
print onward must print "Invoice". The third print
onward must include "Copy of Original" and copy count.