[IMP] open payment matching screen and add domain in payment_id
Before this commit, it was very tedious to return advance because
there was not direct way to match payments once editted or
cancelled from sale order.

This commit opens payment matching screen after sales order is
either editted or cancelled. This commit also adds domains in
payment_id field so that only relevent payments are listed.